+++ CHANGELOG.md
@@ -203,3 +203,4 @@ was removed or the new one was added to folder
 - Dependecies update
 - Logger consumes much less cpu in its own thread
 - New rendering function render shape
+- Sound manager now supports music (mp3, ogg)